+++ Alerting/SmaAlertPackage/AlertClasses/SmaDiskSpaceAlert.py
@@ -1,24 +1,25 @@
-from SmaAlertPackage.SmaAlert import SmaAlert
+from SmaAlertPackage.AlertClasses.SmaAlert import SmaAlert
 from SmaAlertPackage.CommonFunctions.get_pandas_dataframe import get_pandas_dataframe
 from SmaAlertPackage.CommonFunctions.get_pretty_table import get_pretty_table
 from SmaAlertPackage.CommonFunctions.get_sma_params import get_sma_params
 
-class SmaAgDbBackupIssueAlert(SmaAlert):
+class SmaDiskSpaceAlert(SmaAlert):
     '''
-    SYNOPSIS: Class to represent backup issue alert
+    SYNOPSIS: Class to represent disk space alert
     '''
 
-    def __init__(self, alert_key:str=None, alert_owner_team:str='', frequency_minutes:int=10, full_threshold_days:int=8, diff_threshold_hours:int=26, tlog_threshold_minutes:int = 240):
+    def __init__(self, alert_key:str=None, alert_owner_team:str='', frequency_minutes:int=30, disk_warning_pct:float=70, disk_critical_pct:float=85, disk_threshold_gb:int=250, large_disk_threshold_pct:float=92):
         ''' SYNOPSIS: Constructor
         '''
         super().__init__(alert_key, alert_owner_team, frequency_minutes)
-        self.full_threshold_days = full_threshold_days
-        self.diff_threshold_hours = diff_threshold_hours
-        self.tlog_threshold_minutes = tlog_threshold_minutes
+        self.disk_warning_pct = disk_warning_pct
+        self.disk_critical_pct = disk_critical_pct
+        self.disk_threshold_gb = disk_threshold_gb
+        self.large_disk_threshold_pct = large_disk_threshold_pct
         self.alert_pyodbc_resultset = None
 
         self.__df_alert_pyodbc_resultset = None
-        self.__fields_for_display = ["sql_instance", "database_name", "full_latency", "diff_latency", "tlog_latency", "state"]
+        self.__fields_for_display = ["sql_instance", "host_name", "disk_volume", "capacity_mb", "used_pct", "free_mb", "state"]
 
         # severity counts
         self.__critical_count = 0
@@ -128,14 +129,7 @@ class SmaAgDbBackupIssueAlert(SmaAlert):
 
         if self.generate_alert:
             pt = get_pretty_table(self.alert_pyodbc_resultset)
-            #self.description = pt.get_string()
-            if self.__fields_for_display is not None:
-                pt.custom_format = { "full_latency": lambda field, value: self.get_pretty_time(int(value),'days') }
-                pt.custom_format["diff_latency"] = lambda field, value: self.get_pretty_time(value,'hours')
-                pt.custom_format["tlog_latency"] = lambda field, value: self.get_pretty_time(value,'minutes')
-                self.description = pt.get_string(fields=self.__fields_for_display)
-            else:
-                self.description = pt.get_string()
+            self.description = pt.get_string(fields=self.__fields_for_display)
         else:
             self.description = f"Alert cleared."
 
@@ -149,7 +143,7 @@ class SmaAgDbBackupIssueAlert(SmaAlert):
 
     def __compute_sqlmonitor_dashboard_url(self):
         url_grafana_dash = get_sma_params(self.sql_connection, param_key='GrafanaDashboardPortal')[0].param_value
-        url_panel = get_sma_params(self.sql_connection, param_key='url_all_servers_backups_nonag_dbs_dashboard_panel')[0].param_value
+        url_panel = get_sma_params(self.sql_connection, param_key='url_all_servers_disk_utilization_dashboard_panel')[0].param_value
 
         self.__sqlmonitor_dashboard_url = f"{url_grafana_dash}d/{url_panel}"
         if self.verbose:
