+++ Alerting/SmaAlertPackage/AlertClasses/SmaSqlBlockingAlert.py
@@ -1,24 +1,23 @@
-from SmaAlertPackage.SmaAlert import SmaAlert
+from SmaAlertPackage.AlertClasses.SmaAlert import SmaAlert
 from SmaAlertPackage.CommonFunctions.get_pandas_dataframe import get_pandas_dataframe
 from SmaAlertPackage.CommonFunctions.get_pretty_table import get_pretty_table
 from SmaAlertPackage.CommonFunctions.get_sma_params import get_sma_params
 
-class SmaCpuAlert(SmaAlert):
+class SmaSqlBlockingAlert(SmaAlert):
     '''
-    SYNOPSIS: Class to represent cpu alert
+    SYNOPSIS: Class to represent sql blocking alert
     '''
 
-    def __init__(self, alert_key:str=None, alert_owner_team:str='', frequency_minutes:int=10, cpu_warning_pct:float=45, cpu_critical_pct:float=70, average_duration_minutes:int = 5):
+    def __init__(self, alert_key:str=None, alert_owner_team:str='', frequency_minutes:int=15, blocked_counts_threshold:int=1, blocked_duration_max_seconds_threshold:int=120):
         ''' SYNOPSIS: Constructor
         '''
         super().__init__(alert_key, alert_owner_team, frequency_minutes)
-        self.cpu_warning_pct = cpu_warning_pct
-        self.cpu_critical_pct = cpu_critical_pct
-        self.average_duration_minutes = average_duration_minutes
+        self.blocked_counts_threshold = blocked_counts_threshold
+        self.blocked_duration_max_seconds_threshold = blocked_duration_max_seconds_threshold
         self.alert_pyodbc_resultset = None
 
         self.__df_alert_pyodbc_resultset = None
-        self.__fields_for_display = ["sql_instance", "collection_time_latest", "os_cpu_avg", "sql_cpu_avg", "state", "data_points"]
+        self.__fields_for_display = ["sql_instance", "blocked_counts", "blocked_duration_max_seconds", "state", "collection_time"]
 
         # severity counts
         self.__critical_count = 0
@@ -128,6 +127,7 @@ class SmaCpuAlert(SmaAlert):
 
         if self.generate_alert:
             pt = get_pretty_table(self.alert_pyodbc_resultset)
+            pt.custom_format = { "blocked_duration_max_seconds": lambda field, value: self.get_pretty_time(int(value),'seconds') }
             self.description = pt.get_string(fields=self.__fields_for_display)
         else:
             self.description = f"Alert cleared."
