+++ Alerting/SmaAlerts/Alert-DiskSpace.py
@@ -2,25 +2,30 @@ import pyodbc
 import argparse
 from datetime import datetime
 import os
+import sys
+
+# Import parent directory as module to run this script independently
+sys.path.insert(1, os.path.join(sys.path[0], '..'))
 from SmaAlertPackage.CommonFunctions.get_script_logger import get_script_logger
 from SmaAlertPackage.CommonFunctions.connect_dba_instance import connect_dba_instance
 from SmaAlertPackage.CommonFunctions.get_pandas_dataframe import get_pandas_dataframe
 from SmaAlertPackage.CommonFunctions.get_pretty_table import get_pretty_table
-from SmaAlertPackage.CustomFunctions.get_memory_grants_pending import get_memory_grants_pending
-import SmaAlertPackage.SmaMemoryGrantsPendingAlert as sma
+from SmaAlertPackage.CustomFunctions.get_disk_space import get_disk_space
+import SmaAlertPackage.AlertClasses.SmaDiskSpaceAlert as sma
 
 # get Script Name
 script_name = os.path.basename(__file__)
 
-parser = argparse.ArgumentParser(description="Script to raise memory grants pending alert", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+parser = argparse.ArgumentParser(description="Script to raise disk space alert", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 parser.add_argument("--inventory_server", type=str, required=False, action="store", default="localhost", help="Inventory Server")
 parser.add_argument("--inventory_database", type=str, required=False, action="store", default="DBA", help="Inventory Database")
 parser.add_argument("--credential_manager_database", type=str, required=False, action="store", default="DBA", help="Credential Manager Database")
 parser.add_argument("--login_name", type=str, required=False, action="store", default="sa", help="Login name for sql authentication")
 parser.add_argument("--login_password", type=str, required=False, action="store", default="", help="Login password for sql authentication")
-parser.add_argument("--alert_name", type=str, required=False, action="store", default="Alert-MemoryGrantsPending", help="Alert Name")
-parser.add_argument("--alert_job_name", type=str, required=False, action="store", default="(dba) Alert-MemoryGrantsPending", help="Script/Job calling this script")
+parser.add_argument("--alert_name", type=str, required=False, action="store", default="Alert-DiskSpace", help="Alert Name")
+parser.add_argument("--alert_job_name", type=str, required=False, action="store", default="(dba) Alert-DiskSpace", help="Script/Job calling this script")
 parser.add_argument("--alert_owner_team", type=str, required=False, action="store", default="DBA", help="Default team who would own alert")
+#parser.add_argument("--frequency_minutes", type=int, required=False, action="store", default=30, help="Time gap between next execution for same alert")
 parser.add_argument("--verbose", type=bool, required=False, action="store", default=False, help="Extra debug message when enabled")
 parser.add_argument("--log_file", type=str, required=False, action="store", default="", help="Log file path if logging should be done in files.")
 
@@ -38,6 +43,7 @@ if 'Retrieve Parameters' == 'Retrieve Parameters':
     alert_name = args.alert_name
     alert_job_name = args.alert_job_name
     alert_owner_team = args.alert_owner_team
+    #frequency_minutes = args.frequency_minutes
     verbose = args.verbose
     log_file = args.log_file
 
@@ -61,12 +67,14 @@ cursor = cnxn.cursor()
 
 # Create SmaAlert object to retrieve defaults
 logger.info(f"Create SmaAlert child class object with default values..")
-alert_obj = sma.SmaMemoryGrantsPendingAlert()
+alert_obj = sma.SmaDiskSpaceAlert()
 
 if 'Retrieve Class Attribute Defaults' == 'Retrieve Class Attribute Defaults':
     frequency_minutes = alert_obj.frequency_minutes
-    grants_pending_threshold = alert_obj.grants_pending_threshold
-    average_duration_minutes = alert_obj.average_duration_minutes
+    disk_warning_pct = alert_obj.disk_warning_pct
+    disk_critical_pct = alert_obj.disk_critical_pct
+    disk_threshold_gb = alert_obj.disk_threshold_gb
+    large_disk_threshold_pct = alert_obj.large_disk_threshold_pct
 
 # Print variables values
 if 'Print Variables' == 'Print Variables':
@@ -80,19 +88,23 @@ if 'Print Variables' == 'Print Variables':
     logger.info(f"alert_job_name = '{alert_job_name}'")
     logger.info(f"alert_owner_team = '{alert_owner_team}'")
     logger.info(f"frequency_minutes = '{frequency_minutes}'")
-    logger.info(f"grants_pending_threshold = '{grants_pending_threshold}'")
-    logger.info(f"average_duration_minutes = '{average_duration_minutes}'")
+    logger.info(f"disk_warning_pct = '{disk_warning_pct}'")
+    logger.info(f"disk_critical_pct = '{disk_critical_pct}'")
+    logger.info(f"disk_threshold_gb = '{disk_threshold_gb}'")
+    logger.info(f"large_disk_threshold_pct = '{large_disk_threshold_pct}'")
     logger.info(f"verbose = '{verbose}'")
 
 # Get Alert Raw Data
 if 'Get Alert Raw Data' == 'Get Alert Raw Data':
-    logger.info(f"Query table dbo.all_server_volatile_info_history..")
+    logger.info(f"Query table dbo.disk_space_all_servers..")
     query_params = dict(logger = logger,
                         verbose = verbose,
-                        grants_pending_threshold = grants_pending_threshold,
-                        average_duration_minutes = average_duration_minutes
+                        disk_warning_pct = disk_warning_pct,
+                        disk_critical_pct = disk_critical_pct,
+                        disk_threshold_gb = disk_threshold_gb,
+                        large_disk_threshold_pct = large_disk_threshold_pct
                         )
-    alert_pyodbc_resultset = get_memory_grants_pending(cnxn, **query_params)
+    alert_pyodbc_resultset = get_disk_space(cnxn, **query_params)
 
     if len(alert_pyodbc_resultset) > 0:
         logger.info(f"Before creating pt & df on alert_pyodbc_resultset..")
