+++ Alerting/SmaAlerts/Alert-OfflineAgent.py
@@ -2,24 +2,28 @@ import pyodbc
 import argparse
 from datetime import datetime
 import os
+import sys
+
+# Import parent directory as module to run this script independently
+sys.path.insert(1, os.path.join(sys.path[0], '..'))
 from SmaAlertPackage.CommonFunctions.get_script_logger import get_script_logger
 from SmaAlertPackage.CommonFunctions.connect_dba_instance import connect_dba_instance
 from SmaAlertPackage.CommonFunctions.get_pandas_dataframe import get_pandas_dataframe
 from SmaAlertPackage.CommonFunctions.get_pretty_table import get_pretty_table
-from SmaAlertPackage.CustomFunctions.get_nonag_db_backup_issue import get_nonag_db_backup_issue
-import SmaAlertPackage.SmaNonAgDbBackupIssueAlert as sma
+from SmaAlertPackage.CustomFunctions.get_offline_agent import get_offline_agent
+import SmaAlertPackage.AlertClasses.SmaOfflineAgentAlert as sma
 
 # get Script Name
 script_name = os.path.basename(__file__)
 
-parser = argparse.ArgumentParser(description="Script to raise backup issue alert", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+parser = argparse.ArgumentParser(description="Script to raise offline agent alert", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 parser.add_argument("--inventory_server", type=str, required=False, action="store", default="localhost", help="Inventory Server")
 parser.add_argument("--inventory_database", type=str, required=False, action="store", default="DBA", help="Inventory Database")
 parser.add_argument("--credential_manager_database", type=str, required=False, action="store", default="DBA", help="Credential Manager Database")
 parser.add_argument("--login_name", type=str, required=False, action="store", default="sa", help="Login name for sql authentication")
 parser.add_argument("--login_password", type=str, required=False, action="store", default="", help="Login password for sql authentication")
-parser.add_argument("--alert_name", type=str, required=False, action="store", default="Alert-NonAgDbBackupIssue", help="Alert Name")
-parser.add_argument("--alert_job_name", type=str, required=False, action="store", default="(dba) Alert-NonAgDbBackupIssue", help="Script/Job calling this script")
+parser.add_argument("--alert_name", type=str, required=False, action="store", default="Alert-OfflineAgent", help="Alert Name")
+parser.add_argument("--alert_job_name", type=str, required=False, action="store", default="(dba) Alert-OfflineAgent", help="Script/Job calling this script")
 parser.add_argument("--alert_owner_team", type=str, required=False, action="store", default="DBA", help="Default team who would own alert")
 parser.add_argument("--verbose", type=bool, required=False, action="store", default=False, help="Extra debug message when enabled")
 parser.add_argument("--log_file", type=str, required=False, action="store", default="", help="Log file path if logging should be done in files.")
@@ -61,13 +65,10 @@ cursor = cnxn.cursor()
 
 # Create SmaAlert object to retrieve defaults
 logger.info(f"Create SmaAlert child class object with default values..")
-alert_obj = sma.SmaNonAgDbBackupIssueAlert()
+alert_obj = sma.SmaOfflineAgentAlert()
 
 if 'Retrieve Class Attribute Defaults' == 'Retrieve Class Attribute Defaults':
     frequency_minutes = alert_obj.frequency_minutes
-    full_threshold_days = alert_obj.full_threshold_days
-    diff_threshold_hours = alert_obj.diff_threshold_hours
-    tlog_threshold_minutes = alert_obj.tlog_threshold_minutes
 
 # Print variables values
 if 'Print Variables' == 'Print Variables':
@@ -81,21 +82,15 @@ if 'Print Variables' == 'Print Variables':
     logger.info(f"alert_job_name = '{alert_job_name}'")
     logger.info(f"alert_owner_team = '{alert_owner_team}'")
     logger.info(f"frequency_minutes = '{frequency_minutes}'")
-    logger.info(f"full_threshold_days = '{full_threshold_days}'")
-    logger.info(f"diff_threshold_hours = '{diff_threshold_hours}'")
-    logger.info(f"tlog_threshold_minutes = '{tlog_threshold_minutes}'")
     logger.info(f"verbose = '{verbose}'")
 
 # Get Alert Raw Data
 if 'Get Alert Raw Data' == 'Get Alert Raw Data':
-    logger.info(f"Query table dbo.all_server_volatile_info_history..")
+    logger.info(f"Query table dbo.services_all_servers..")
     query_params = dict(logger = logger,
-                        verbose = verbose,
-                        full_threshold_days = full_threshold_days,
-                        diff_threshold_hours = diff_threshold_hours,
-                        tlog_threshold_minutes = tlog_threshold_minutes
+                        verbose = verbose
                         )
-    alert_pyodbc_resultset = get_nonag_db_backup_issue(cnxn, **query_params)
+    alert_pyodbc_resultset = get_offline_agent(cnxn, **query_params)
 
     if len(alert_pyodbc_resultset) > 0:
         logger.info(f"Before creating pt & df on alert_pyodbc_resultset..")
