+++ Alerting/SmaAlerts/Alert-Tempdb.py
@@ -2,26 +2,29 @@ import pyodbc
 import argparse
 from datetime import datetime
 import os
+import sys
+
+# Import parent directory as module to run this script independently
+sys.path.insert(1, os.path.join(sys.path[0], '..'))
 from SmaAlertPackage.CommonFunctions.get_script_logger import get_script_logger
 from SmaAlertPackage.CommonFunctions.connect_dba_instance import connect_dba_instance
 from SmaAlertPackage.CommonFunctions.get_pandas_dataframe import get_pandas_dataframe
 from SmaAlertPackage.CommonFunctions.get_pretty_table import get_pretty_table
-from SmaAlertPackage.CustomFunctions.get_offline_server import get_offline_server
-import SmaAlertPackage.SmaOfflineServerAlert as sma
+from SmaAlertPackage.CustomFunctions.get_tempdb import get_tempdb
+import SmaAlertPackage.AlertClasses.SmaTempdbAlert as sma
 
 # get Script Name
 script_name = os.path.basename(__file__)
 
-parser = argparse.ArgumentParser(description="Script to raise offline server alert", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+parser = argparse.ArgumentParser(description="Script to raise tempdb alert", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 parser.add_argument("--inventory_server", type=str, required=False, action="store", default="localhost", help="Inventory Server")
 parser.add_argument("--inventory_database", type=str, required=False, action="store", default="DBA", help="Inventory Database")
 parser.add_argument("--credential_manager_database", type=str, required=False, action="store", default="DBA", help="Credential Manager Database")
 parser.add_argument("--login_name", type=str, required=False, action="store", default="sa", help="Login name for sql authentication")
 parser.add_argument("--login_password", type=str, required=False, action="store", default="", help="Login password for sql authentication")
-parser.add_argument("--alert_name", type=str, required=False, action="store", default="Alert-OfflineServer", help="Alert Name")
-parser.add_argument("--alert_job_name", type=str, required=False, action="store", default="(dba) Alert-OfflineServer", help="Script/Job calling this script")
+parser.add_argument("--alert_name", type=str, required=False, action="store", default="Alert-Tempdb", help="Alert Name")
+parser.add_argument("--alert_job_name", type=str, required=False, action="store", default="(dba) Alert-Tempdb", help="Script/Job calling this script")
 parser.add_argument("--alert_owner_team", type=str, required=False, action="store", default="DBA", help="Default team who would own alert")
-#parser.add_argument("--frequency_minutes", type=int, required=False, action="store", default=30, help="Time gap between next execution for same alert")
 parser.add_argument("--verbose", type=bool, required=False, action="store", default=False, help="Extra debug message when enabled")
 parser.add_argument("--log_file", type=str, required=False, action="store", default="", help="Log file path if logging should be done in files.")
 
@@ -39,7 +42,6 @@ if 'Retrieve Parameters' == 'Retrieve Parameters':
     alert_name = args.alert_name
     alert_job_name = args.alert_job_name
     alert_owner_team = args.alert_owner_team
-    #frequency_minutes = args.frequency_minutes
     verbose = args.verbose
     log_file = args.log_file
 
@@ -63,10 +65,13 @@ cursor = cnxn.cursor()
 
 # Create SmaAlert object to retrieve defaults
 logger.info(f"Create SmaAlert child class object with default values..")
-alert_obj = sma.SmaOfflineServerAlert()
+alert_obj = sma.SmaTempdbAlert()
 
 if 'Retrieve Class Attribute Defaults' == 'Retrieve Class Attribute Defaults':
     frequency_minutes = alert_obj.frequency_minutes
+    data_used_warning_pct = alert_obj.data_used_warning_pct
+    data_used_critical_pct = alert_obj.data_used_critical_pct
+    data_used_threshold_gb = alert_obj.data_used_threshold_gb
 
 # Print variables values
 if 'Print Variables' == 'Print Variables':
@@ -80,15 +85,21 @@ if 'Print Variables' == 'Print Variables':
     logger.info(f"alert_job_name = '{alert_job_name}'")
     logger.info(f"alert_owner_team = '{alert_owner_team}'")
     logger.info(f"frequency_minutes = '{frequency_minutes}'")
+    logger.info(f"data_used_warning_pct = '{data_used_warning_pct}'")
+    logger.info(f"data_used_critical_pct = '{data_used_critical_pct}'")
+    logger.info(f"data_used_threshold_gb = '{data_used_threshold_gb}'")
     logger.info(f"verbose = '{verbose}'")
 
 # Get Alert Raw Data
 if 'Get Alert Raw Data' == 'Get Alert Raw Data':
-    logger.info(f"Query table dbo.instance_details..")
+    logger.info(f"Query table dbo.tempdb_consumers_all_servers..")
     query_params = dict(logger = logger,
-                        verbose = verbose
+                        verbose = verbose,
+                        data_used_warning_pct = data_used_warning_pct,
+                        data_used_critical_pct = data_used_critical_pct,
+                        data_used_threshold_gb = data_used_threshold_gb
                         )
-    alert_pyodbc_resultset = get_offline_server(cnxn, **query_params)
+    alert_pyodbc_resultset = get_tempdb(cnxn, **query_params)
 
     if len(alert_pyodbc_resultset) > 0:
         logger.info(f"Before creating pt & df on alert_pyodbc_resultset..")
